Add extension of global and local config hashmaps
The extension with `global_config` on the right side means that means
that the values that are present on the `local_config` will be the ones
that have preference, ending in values like the following:

```
{
    "commands": {
        "run": "run globally",
        "test": "cargo test -- --nocapture"
    }
}
```

Let's imagine that we have a cargo project, and have added a local
dyncomp.json file in the main directory of it. "run" is a key that is
defined globally, hence running `dyncomp run` will run "run
globally". Since we added whe "test" key locally, when using `dyncomp
test` we'll run "cargo test -- --nocapture".